services/requirements.txt
Flask
mysql-connector
wtforms
